% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fulltest.R
\name{buildFromCsv}
\alias{buildFromCsv}
\title{Build Predictive Models from CSV}
\usage{
buildFromCsv(csv, diags, modelType = "SVM", multiLabMod = "!d",
  minCodSiz = 0)
}
\arguments{
\item{csv}{a data frame with 2 columns: 'NDA' (a unique line identifier), and 'TEXT (free text).}

\item{diags}{a data frame with 2 columns: 'NDA' (a unique line identifier), and 'DIAG' (alpha-num codes).}

\item{modelType}{model type from \{"SVM", "NB"\} for Support Vector Machine, and Naive Bayes respectively.}

\item{multiLabMod}{a string argument to specify how to deal with multilabeled texts. Put "d" if you want to duplicate them (the same text is considered multiple times with one code at each), or any other value to ignore multilabeled texts.}

\item{minCodSiz}{minimum code size to be reached (numeric). If this argument is set to K, then the codes with less than K texts will be bootstrapped to reach this minimum.}
}
\value{
an object of type "svm" or "nb" (depends on the specified 'modelType' argument.).
If you want to use the built model within 'mlcodage' Web service, you must add it to the 'data' folder of the package in the form of an RData object with extension ".model.RData', then re-compile the package.
}
\description{
Build a new code prediction model from CSV data. The function takes two mandatory 2-column dataframes: 'csv' and 'diags'. See below for details.
}
\examples{
csv = data.frame(NDA=c("T1", "T2"), TEXT=c("text numer one", "text number two"))
diags = data.frame(NDA=c("T1", "T2"), DIAG=c("CODE1", "CODE2"))
buildFromCsv(csv, diags, modelType="SVM", multiLabMod="!d", minCodSiz=0)
}

