% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocpu.R
\name{getAvailModels}
\alias{getAvailModels}
\title{Get Available Models}
\usage{
getAvailModels()
}
\value{
character vector with the names of available models.
}
\description{
Returns the list of all available models that can be used for code prediction. The models are stored as 'RData' files within the 'data' folder of the package.
If you want to use custom models, please add them manually to the 'data' folder (with extension '.model.RData') then re-build the package.
}
\examples{
getAvailModels() # returns c("Chirurgie-generale", "Chirurgie-maxillo-faciale", "Chirurgie-plastique-Pr-Revol", "Urologie")
}

