% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocpu.R
\name{ccamCode}
\alias{ccamCode}
\title{CCAM Code Prediction}
\usage{
ccamCode(texte = "", service = "", nbCodes = 10)
}
\arguments{
\item{service}{service name, from the list of built-in models : \{'urologie', 'chirgen', 'chirplas', 'chirmaxfac'\}. 
This parameter is used to pick the right model for code prediction.}

\item{nbCodes}{the desired number of returned codes (default 10).}

\item{text}{medical text to be encoded.}
}
\value{
string with JSON data with the ordered list of predicted codes. Each is assigned with code label and probability.
If an error occurs, the function returns a string with error description.
}
\description{
CCAM code prediction function. It is used by OpenCPU Web service for Web-based prediction and can be used directly from R console.
}
\examples{
ccamCode("Résection de tumeurs de la vessie, par endoscopie", "urologie", 5) # returns a JSON string with predicted codes and probabilities.
}

